Code cleaning
